+++ src/Phinx/Db/Table/Column.php
@@ -66,7 +66,7 @@ class Column
     protected $type;
 
     /**
-     * @var float|null
+     * @var int|null
      */
     protected $limit;
 
@@ -199,10 +199,10 @@ class Column
     /**
      * Sets the column limit.
      *
-     * @param float|null $limit Limit
+     * @param int|null $limit Limit
      * @return $this
      */
-    public function setLimit(?float $limit)
+    public function setLimit(?int $limit)
     {
         $this->limit = $limit;
 
@@ -212,9 +212,9 @@ class Column
     /**
      * Gets the column limit.
      *
-     * @return float|null
+     * @return int|null
      */
-    public function getLimit(): ?float
+    public function getLimit(): ?int
     {
         return $this->limit;
     }
@@ -360,10 +360,10 @@ class Column
      * For example `DECIMAL(5,2)`, 5 is the precision and 2 is the scale,
      * and the column could store value from -999.99 to 999.99.
      *
-     * @param float|null $precision Number precision
+     * @param int|null $precision Number precision
      * @return $this
      */
-    public function setPrecision(?float $precision)
+    public function setPrecision(?int $precision)
     {
         $this->setLimit($precision);
 
@@ -376,9 +376,9 @@ class Column
      * For example `DECIMAL(5,2)`, 5 is the precision and 2 is the scale,
      * and the column could store value from -999.99 to 999.99.
      *
-     * @return float
+     * @return int|null
      */
-    public function getPrecision(): ?float
+    public function getPrecision(): ?int
     {
         return $this->limit;
     }
